==> 02_Session.Rmd <==
---
title: "Session 2 notes"
author: "Tim Riffe"
date: "2025-06-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Life tables

For today's practice, we will read in the HMD data
```{r}
library(tidyverse)
hmd <- read_csv("https://github.com/timriffe/KOSTAT_Workshop1/raw/refs/heads/master/Data/hmd.csv.gz")

# we have 9352 lifetables
hmd |> 
  select(country, sex, year) |> 
  distinct() |> 
  nrow()

hmd |> 
  pull(sex) |> 
  unique()
```

# step 0 visualize it

```{r}
hmd |> 
  filter(year >= 1950,
         !country %in% c("Iceland","Luxembourg")) |> 
  mutate(mx = if_else(mx == 0, NA, mx)) |> 
  ggplot(mapping = aes(x = age, y = mx, group = year))+
  geom_line(alpha = .03) +
  scale_y_log10() +
  facet_wrap(~sex) 
```

Looks at one:

```{r}
hmd$country |> unique()
hmd |> 
  filter(country == "Sweden",
         year == 2015) |> 
  ggplot(mapping = aes(x = age, y =mx, color = sex)) +
  geom_line() +
  scale_y_log10()
```


# step 1: convert rates to probabilties

For single-age data, the formula to convert rates $m_x$ to probabilities $q_x$ is:

$$ q_x = \frac{m_x}{(1 + (1-a_x) \cdot m_x)}$$
 The formula still works for wider age groups ($n$ years wide) with a small modification:
 
 $$ {}_nq_x = \frac{n \cdot {}_nm_x}{1 + (n- {}_na_x)\cdot{}_nm_x}$$
Note: $n$ on the left subscript is to indicate that the variable covers an $n$-year wide interval starting at age $x$.

Select a subset-- what countries are available to pick from?
```{r}
hmd |> 
  pull(country) |> 
  unique()
```

now calculate qx:
```{r}
es <- 
  hmd |> 
  filter(country == "Spain",
         sex == "f",
         year == 2019) 
es |> 
  mutate(qx = mx / (1 + (1 - ax) * mx),
         qx = if_else(qx > 1, 1, qx), # optional
         qx = if_else(age == max(age), 1, qx)) # convention
```

Now let's put that in a function:

```{r}
mx_to_qx <- function(mx, ax){
  qx              <- mx / (1 + (1 - ax) * mx)
  # make sure it obeys the rules!
  qx[qx > 1]      <- 1
  # no one gets out!
  qx[length(qx)]  <- 1
  return(qx)
}
es <- 
es |> 
  mutate(qx = mx_to_qx(mx, ax)) 
```

# now calculate survivorship

survival probabilities are the complement of death probabilities...
$$ p_x  = 1 - q_x $$
These are the probabilities can chain together into overall survivorship:
$$ \ell_x = \prod _{0}^{x-1}p_x$$

```{r}
qx_to_lx <- function(qx, radix = 1e5){
  n <- length(qx)
  qx <- c(0,qx[-n])
  lx <- radix * cumprod(1-qx)
  return(lx)
}
```
Now let's try it out:

```{r}
es |> 
  mutate(qx = mx_to_qx(mx, ax),
         lx = qx_to_lx(qx),
         dx = qx * lx) |> 
  ggplot(aes(x = age, y = dx)) +
  geom_line()
```

# Exercise: Approximate lifetable exposure

$$ L_x = \ell_x - (d_x \cdot(1 - a_x))$$

anatomy of a function
```{r, eval = FALSE}
fun_name <- function(args, more_args){
  # do stuff with args
  return(result)
}
```

my take:
```{r}
calc_Lx <- function(lx,dx,ax){
  n     <- length(lx)
  Lx    <- lx - (dx * (1 - ax))
  Lx[n] <- lx[n] * ax[n]
  return(Lx)
}
```

apply it:

```{r}

es |> 
  mutate(qx = mx_to_qx(mx = mx, ax = ax),
         lx = qx_to_lx(qx, radix = 1),
         dx = lx * qx,
         Lx = calc_Lx(lx = lx, dx = dx, ax = ax))
```

# Calculate remaining life expectancy

```{r}
calc_ex <- function(Lx, lx){
  # this one is cumbersome:
  # Tx <- sum(Lx) - c(0,cumsum(Lx[-111]))  
  # this one makes your head spin:
  Tx <- Lx |> rev() |> cumsum() |> rev()
  ex <- Tx / lx # condition it! (scale it up)
  return(ex)
}
```

# Exercise: apply all steps to calculate a full lifetable

This works by setting up a complete `mutate()` call for a lifetable (all steps), then insert a `group_by()` statement.
```{r, eval = FALSE}
hmd |>
  group_by(country, sex, year) |> 
  mutate(qx = mx_to_qx(mx=mx,ax),
         lx = qx_to_lx(qx = qx, radix = 1e5),
         dx = qx * lx,
         Lx = calc_Lx(lx = lx, dx = dx, ax = ax),
         ex = calc_ex(Lx = Lx,lx = lx))
```


What might a full lifetable function look like?

```{r}
calc_LT <- function(mx, ax, age, radix = 1e5){
  lt <- tibble(age, mx, ax) |> 
          mutate(qx = mx_to_qx(mx = mx, ax = ax),
                 lx = qx_to_lx(qx = qx, radix = radix),
                 dx = qx * lx,
                 Lx = calc_Lx(lx = lx, dx = dx, ax = ax),
                 ex = calc_ex(Lx = Lx,lx = lx))
  return(lt)
} 
```

How to load a bunch of your own functions

```{r}
source("https://raw.githubusercontent.com/timriffe/KOSTAT_Workshop1/refs/heads/master/lifetable_functions2025.R")
```
# Exercises-
1. Make a scatterplot (points), using `geom_point()` of:
$m_0$ vs $e_0$ (infant mortality vs life expectancy). Maybe log infant mortality scale. That means: calculate the full lifetables, then subset to `age == 0`, then make the plot. sex can be colors

```{r}
hmd_lt <-
  hmd |> 
  group_by(country, sex, year) |> 
  reframe(calc_LT(mx = mx, ax = ax, age = age, radix = 1e5)) 

hmd_lt |> 
  ungroup() |> 
  filter(age == 0, year > 1850) |> 
  ggplot(mapping = aes(x = ex, y = mx, color = year)) +
  geom_point(alpha = .3) +
  geom_smooth(method = "loess", color = "red")+
  scale_y_log10() +
  facet_wrap(~sex) +
  theme_minimal()
```

2. Make a function that calculates the standard deviation (at birth) of age at death. That is:
It takes dx, ax, and age as arguments, and does this:
$$ var_0 = \sum_0^\infty(x - e_0)^2 \cdot d_x/\ell_0$$
$$ \sigma_0 = \sqrt(var_x)$$
The function will only return one value, meaning you'll need to apply it inside `summarize()`, and also save $e_0$ inside that summarize (`e0 = ex[age == 0]`)

```{r}
calc_sd0 <- function(dx, ax, age){
  avg_age_at_death_in_interval <- age + ax
  dx  <- dx / sum(dx)
  e0  <- sum(avg_age_at_death_in_interval * dx)
  v0  <- sum((avg_age_at_death_in_interval - e0)^2 * dx)
  sd0 <- sqrt(v0)
  return(sd0)
}
```


Now make a scatterplot of the standard deviation (at birth) and life expectancy at birth, with sex mapped to color (or in a facet).

```{r}
hmd_lt |> 
  filter(year > 1850) |> 
  group_by(country, sex, year) |> 
  summarize(e0 = ex[age==0],
            sd0 = calc_sd0(dx = dx, ax = ax, age = age)) |> 
  ggplot(aes(x = e0, y = sd0, color = year)) +
  geom_point(alpha = .3) +
  geom_smooth(color = "red") +
  theme_minimal() +
  facet_wrap(~sex)
```








```{r}
hmd |> 
  group_by(country, sex, year) |> 
  reframe(calc_LT(mx = mx, ax = ax, age = age))
```

# Exercise

Make a scatterplot of adult mortality against infant or child mortality.

$$ {}_{45}q_{15}$$

$$ {}_{45}q_{15} = \frac{\sum _{15}^{59} d_x}{\ell_{15}}$$

$$ {}_{45}q_{15} = 1 - (\frac{\ell_{60}}{\ell_{15}})$$

```{r}
calc_adult_mort <- function(lx, age){
  l60        <- lx[age == 60]
  l15        <- lx[age == 15]
  adult_mort <- 1 - l60 / l15
  return(adult_mort)
}
```

```{r}
hmd_lt |> 
  group_by(country, sex, year) |> 
  summarize(m0 = mx[age == 0],
            adult_mort = calc_adult_mort(lx = lx, age = age),
            .groups = "drop") |> 
  ggplot(mapping = aes(x = m0, y = adult_mort)) +
  geom_point(alpha = .3) +
  scale_x_log10()
```
