==> 04_session.Rmd <==
---
title: "session 4 notes"
author: "Tim Riffe"
date: "2025-06-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#

```{r}

a1 <- .00008
a2 <- .0003
b <-.07
age <- 30:100
gompmx <- function(a,b,x){
  a * exp(b * x)
}
mx_to_lx_cheap <- function(mx){
  n <- length(mx)
  c(1,exp(-cumsum(mx[-n])))
}
mx1 <- gompmx(a1,b,age)
mx2 <- gompmx(a2,b,age)
lx1 <- mx_to_lx_cheap(mx1)
lx2 <- mx_to_lx_cheap(mx2)


lx_tidy <- tibble(age, lx1, lx2)
prev30 <- c(rep(.01,70), rep(.1,7), rep(.2,7), rep(.3,7), rep(.4,70))

cohorts <- seq_along(prev30) + 1900
cohorts_tidy <- tibble(cohort = cohorts, prev_high = prev30)
library(tidyverse)

lx_data <- cross_join(cohorts_tidy,lx_tidy) 
prev_data <- 
lx_data |> 
  mutate(prev1 = prev_high * lx1,
         prev2 = (1 - prev_high) * lx2,
         year = cohort + age) |> 
  select(year, age, prev1, prev2) |> 
  group_by(year) |> 
  mutate(n = n()) |> 
  ungroup() |> 
  filter(n == 71) |> 
  select(-n)

prev_data |> 
  ggplot(aes(x = year, y = age, fill = prev1)) +
  geom_tile() +
  coord_equal()


```

The above code was shared via a script to be sourced from grithub.

# load lifetable functions
```{r}
# gets the lifetable function
source("https://raw.githubusercontent.com/timriffe/KOSTAT_Workshop1/refs/heads/master/lifetable_functions2025.R")

# run the prevalence simulation
source("https://raw.githubusercontent.com/timriffe/KOSTAT_Workshop1/refs/heads/master/04_simulation.R")

```

```{r}
mx_national <-  
  # collect edu-specific mx in table
  tibble(age = age, mx1 = mx1, mx2 = mx2) |> 
  # join to each year of edu prevalence
  right_join(prev_data, by = join_by(age)) |> 
  arrange(year, age) |> 
  # calculate national average mx
  mutate(mx = mx1 * prev1 + mx2 * prev2) |> 
  select(year, age, mx) |> 
  mutate(ax = .5)
```

# Exercise:
1. Calculate lifetables for each year (use `calc_LT()` inside of `reframe()`)
2. filter down to age 30
3. Make a graph of remaining life expectancy at age 30.

```{r}
mx_national |> 
  group_by(year) |> 
  reframe(calc_LT(mx = mx, ax = ax, age = age)) |> 
  filter(age == 30) |> 
  ggplot(mapping = aes(x = year, y = ex + 30)) +
  geom_line() +
  theme_minimal()
```

# load wpp mortality rates

```{r}
mxwpp <- read_csv("https://github.com/timriffe/KOSTAT_Workshop1/raw/refs/heads/master/Data/mx1dt.csv.gz")

```

# append country names

```{r}
library(countrycode)
mxwpp <-
  mxwpp |> 
  mutate(iso3 = countrycode(sourcevar = country_code,
                            origin = "un",
                            destination = "iso3c")) |> 
  filter(!is.na(iso3))
```

# reshape data to long format

```{r}
mxwpp <- 
  mxwpp |> 
  pivot_longer(starts_with("mx"), 
               names_to = "sex",
               values_to = "mx") |> 
  mutate(sex = case_when(sex == "mxM" ~ "male",
                         sex == "mxF" ~ "female",
                         TRUE ~ "total")) |> 
  arrange(iso3, year, sex, age)
```

# figure out infant a0

```{r}
mxwpp <-
  mxwpp |> 
  filter(sex != "total") |> 
  mutate(ax = case_when(
    sex == "male" & age == 0 & between(mx, 0, 0.02300) ~ 0.14929 - 1.99545 * mx,
    sex == "male" & age == 0 & between(mx, 0.02300001, 0.08307) ~ 0.02832 + 3.26021 * mx,
    sex == "male" & age == 0 & between(mx, 0.08307001, 1) ~ 0.29915,
    sex == "female" & age == 0 & between(mx, 0, 0.01724) ~ 0.14903 - 2.05527 * mx,
    sex == "female" & age == 0 & between(mx, 0.01724001, 0.06891) ~ 0.04667 + 3.88089 * mx,
    sex == "female" & age == 0 & between(mx, 0.06891001,1) ~ 0.31411,
    age == max(age) ~ 1 / mx, # constant mortality for closeout (the mean survival time)
    TRUE ~ .5                 # if subject to constant mortality is the inverse of the rate
  ))
```

# Exercise: calculate all lifetables.

For very fast calculations you can use this approximation for life expectancy at birth; this returns a scalar (a single value), so you should run this inside `summarize()`.

$$ \ell(x) = e^{-\int_0^x m(t) dt}$$

$$ \ell(x) = 1 - F(x) = e^{-H(x)}$$


```{r}
mx_to_e0_cheap <- function(mx){
  lx <- c(1,exp(-cumsum(mx)))
  sum(lx) - .5
}
e0_fast <-
  mxwpp |> 
  group_by(iso3, country_code, name, sex, year) |> 
  summarize(e0 = mx_to_e0_cheap(mx), .groups = "drop") 
```

Here's the rigorous way:

```{r, eval = FALSE}
mxwpp |> 
  group_by(iso3, country_code, name, sex, year) |> 
  reframe(calc_LT(mx = mx, ax = ax, age = age)) |> 
  filter(age == 0)
```

# Exercise: add continent names

Add continent names to a new column using the `countrycode` package.

```{r}
e0_fast <-
  e0_fast |> 
  mutate(continent = countrycode(iso3, 
                                 origin = "iso3c", 
                                 destination = "continent"))
```


# Make exploratory plot

Make $e(0)$ trajectories as lines? Each country is a line. And the colors of the lines can come from `continent`. You'll want to use `alplha = __` inside `geom_line()`.

```{r}
e0_fast |> 
  ggplot(mapping = aes(x = year, 
                       y = e0, 
                       color = continent, 
                       groups = iso3)) +
  geom_point(alpha = .5) +
  facet_wrap(~sex)
```

# highlight a country on top the cloud

```{r}
e0_fast |> 
  ggplot(mapping = aes(x = year,
                       y = e0)) +
  geom_point(alpha = .05) +
  geom_line(data = e0_fast |> filter(iso3 == "KOR"),
            color = "red", linewidth = 2) +
  facet_wrap(~sex)
```

# Calculate speed of improvement
By continents and decades 1950-2010.

Here, use this handy help function to calculate slopes!!
```{r}
calc_slope <- function(year, e0){
  coefs <- lm(e0~year) |> coef()
  coefs[["year"]]
}
```

Trick to figure out decades:
The line `years - years %% 10` tells you which decade each year falls within; so you can make a decade column inside `mutate()` using that as the trick.
```{r}
years <- 1980:2020
decade <- years - years %% 10
tibble(year = years, decade)
```

Now you're ready to group by continent (or something else) and decade and use `calc_slope()` inside `summarize()` to tell me the speed of improvement.

```{r}
e0_fast <- 
  e0_fast |> 
  filter(year < 2020) |> 
  mutate(decade = year - year %% 10)
```

```{r}
e0_fast |> 
  group_by(continent, decade) |> 
  summarize(pace = calc_slope(year = year, 
                              e0 = e0)) |> 
  mutate(pace = round(pace, 2)) |> 
  pivot_wider(names_from = decade,
              values_from = pace)
```

# New exercise
Within each year, pick out the highest life expectancy at birth. This is a `filter()` line.

Inside `summarize()` use `calc_slope()`.

```{r}
e0_fast |> 
  ungroup() |> 
  filter(year < 2020) |> 
  group_by(year) |> 
  filter(e0 == max(e0)) |> 
  ggplot(mapping = aes(x = year, y = e0)) +
  geom_point(mapping = aes(color = name)) +
  geom_smooth(method = "lm")

e0_fast |> 
  ungroup() |> 
  filter(year < 2020) |> 
  group_by(year) |> 
  filter(e0 == max(e0)) |> 
  ungroup() |> 
  summarize(pace = calc_slope(year = year, e0 = e0))
```

# 

```{r}
mxwpp |> 
  filter(year < 2020,
         iso3 == "AUS",
         sex == "female") |> 
  ggplot(mapping = aes(x = year, y = age, z = log(mx))) +
  coord_equal() +
  geom_contour_filled()
```


```{r}

```
